+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/DefaultTextFieldBuilder.java
@@ -20,21 +20,15 @@ package is.codion.swing.common.ui.component.text;
 
 import is.codion.common.value.Value;
 import is.codion.swing.common.ui.component.value.ComponentValue;
-import is.codion.swing.common.ui.dialog.ListSelectionDialogBuilder.SingleSelector;
-import is.codion.swing.common.ui.key.KeyEvents;
 
-import javax.swing.AbstractAction;
 import javax.swing.Action;
 import javax.swing.JTextField;
 import javax.swing.SwingConstants;
-import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.text.Format;
 
 import static is.codion.common.Text.nullOrEmpty;
 import static is.codion.swing.common.ui.component.text.SizedDocument.sizedDocument;
-import static java.awt.event.InputEvent.CTRL_DOWN_MASK;
-import static java.awt.event.KeyEvent.VK_SPACE;
 import static java.util.Objects.requireNonNull;
 
 class DefaultTextFieldBuilder<T, C extends JTextField, B extends TextFieldBuilder<T, C, B>> extends AbstractTextComponentBuilder<T, C, B>
@@ -45,7 +39,6 @@ class DefaultTextFieldBuilder<T, C extends JTextField, B extends TextFieldBuilde
 	private int columns = -1;
 	private Action action;
 	private ActionListener actionListener;
-	private SingleSelector<T> selector;
 	private Format format;
 	private int horizontalAlignment = SwingConstants.LEADING;
 	private String hint;
@@ -76,12 +69,6 @@ class DefaultTextFieldBuilder<T, C extends JTextField, B extends TextFieldBuilde
 		this.actionListener = actionListener;
 
 		return transferFocusOnEnter(false);
-	}
-
-	@Override
-	public final B selector(SingleSelector<T> selector) {
-		this.selector = requireNonNull(selector);
-		return self();
 	}
 
 	@Override
@@ -118,9 +105,6 @@ class DefaultTextFieldBuilder<T, C extends JTextField, B extends TextFieldBuilde
 		if (actionListener != null) {
 			textField.addActionListener(actionListener);
 		}
-		if (selector != null) {
-			setSelector(textField, selector);
-		}
 		if (hint != null && textField instanceof HintTextField) {
 			((HintTextField) textField).hint().set(hint);
 		}
@@ -147,30 +131,5 @@ class DefaultTextFieldBuilder<T, C extends JTextField, B extends TextFieldBuilde
 
 	protected final Format format() {
 		return format;
-	}
-
-	private void setSelector(C textField, SingleSelector<T> selector) {
-		KeyEvents.builder(VK_SPACE)
-						.modifiers(CTRL_DOWN_MASK)
-						.action(new SelectionAction<>(textField, selector))
-						.enable(textField);
-	}
-
-	private static final class SelectionAction<T> extends AbstractAction {
-
-		private final JTextField textField;
-		private final SingleSelector<T> selector;
-
-		private SelectionAction(JTextField textField, SingleSelector<T> selector) {
-			super("DefaultTextFieldBuilder.SelectionAction");
-			this.textField = textField;
-			this.selector = selector;
-		}
-
-		@Override
-		public void actionPerformed(ActionEvent e) {
-			selector.select(textField)
-							.ifPresent(value -> textField.setText(value.toString()));
-		}
 	}
 }