+++ changelog.md
@@ -29,6 +29,9 @@ Codion Change Log
 - DefaultFileSelectionDialogBuilder.selectFileToSave() and selectFilesOrDirectories() bug fixed, incorrect class used for synchronization.
 - SelectionDialogBuilder split into ListSelectionDialogBuilder and ComboBoxSelectionDialogBuilder.
 - Dialogs.selectionDialog() renamed listSelectionDialog() and comboBoxSelectionDialog() added.
+- ListSelectionDialogBuilder.SingleSelector and MultiSelector removed.
+- Dialogs.singleSelector() and multiSelector() removed.
+- TextFieldBuilder.selector() removed.
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel.Builder.applicationModelFactory() and applicationPanelFactory() renamed applicationModel() and applicationPanel().
 - EntityApplicationPanel.createLogControls() now protected final.